requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
blinker==1.5
boto3==1.26.94
botocore==1.29.94
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
Faker==17.6.0
favicon==0.7.0
filelock==3.10.0
fonttools==4.39.2
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
htbuilder==0.6.1
huggingface-hub==0.13.2
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.2.0
markdownlit==0.0.5
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
numpy==1.24.2
openai==0.27.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
pymdown-extensions==9.10
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.3.2
s3transfer==0.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4
st-annotated-text==3.0.0
streamlit==1.20.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-embedcode==0.1.2
streamlit-extras==0.2.6
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.11.1
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==2.0.0
tornado==6.2
tqdm==4.65.0
transformers==4.27.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==2.3.1
yarl==1.8.2
zipp==3.15.0
